Template toolkit file for history
I wrote this file a long time ago, after searching for hours how to add
plugins to template toolkit. In the end, I went for something else. But
I don't want to completely lose the idea. Let's commit it here, and
remove it right away, so that it stays somewhere in the history of this
repo.

Signed-off-by: Joël Porquet <joel@porquet.org>